requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
amqp==5.3.1
    # via kombu
anyio==4.10.0
    # via httpx
asgiref==3.9.1
    # via django
billiard==4.2.1
    # via celery
boto3==1.40.21
    # via gitsap (pyproject.toml)
botocore==1.40.21
    # via
    #   boto3
    #   s3transfer
celery==5.5.3
    # via gitsap (pyproject.toml)
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via pygit2
click==8.2.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
django==5.2.5
    # via
    #   gitsap (pyproject.toml)
    #   djangorestframework
djangorestframework==3.16.1
    # via gitsap (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via gitsap (pyproject.toml)
idna==3.10
    # via
    #   anyio
    #   httpx
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.5.4
    # via celery
packaging==25.0
    # via kombu
pillow==11.3.0
    # via gitsap (pyproject.toml)
prompt-toolkit==3.0.52
    # via click-repl
pycparser==2.22
    # via cffi
pygit2==1.18.2
    # via gitsap (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
python-dotenv==1.1.1
    # via gitsap (pyproject.toml)
redis==6.4.0
    # via gitsap (pyproject.toml)
s3transfer==0.13.1
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlparse==0.5.3
    # via django
typing-extensions==4.15.0
    # via anyio
tzdata==2025.2
    # via kombu
urllib3==2.5.0
    # via botocore
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
